move devtools caller-location into browser-compat subfolder and prune helper tests
